Generalize connection command running and I/O for all connection types
